Moved testConsumerOffsetPath from ZkDataTest to KafkaZkClientTest
