+++ core/src/test/scala/unit/kafka/zk/KafkaZkClientTest.scala
@@ -1043,4 +1043,20 @@ class KafkaZkClientTest extends ZooKeeperTestHarness {
     assertTrue(zkClient.deleteDelegationToken(tokenId))
     assertEquals(None, zkClient.getDelegationTokenInfo(tokenId))
   }
+
+  @Test
+  def testConsumerOffsetPath(): Unit = {
+    def getConsumersOffsetsZkPath(consumerGroup: String, topic: String, partition: Int): String = {
+      s"/consumers/$consumerGroup/offsets/$topic/$partition"
+    }
+
+    val consumerGroup = "test-group"
+    val topic = "test-topic"
+    val partition = 2
+
+    val expectedConsumerGroupOffsetsPath = getConsumersOffsetsZkPath(consumerGroup, topic, partition)
+    val actualConsumerGroupOffsetsPath = ConsumerOffset.path(consumerGroup, topic, partition)
+
+    assertEquals(expectedConsumerGroupOffsetsPath, actualConsumerGroupOffsetsPath)
+  }
 }
